Modify: 회원가입 & 로그인 컴포넌트 분리 및 도메인에 따른 페이지 구현 완료
